+++ genienlp/tasks/almond_task.py
@@ -28,10 +28,11 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 import logging
 import os
-import re
 from collections import defaultdict
 
 import torch
+
+from genienlp.data_utils.almond_utils import split_text_into_sentences
 
 from ..data_utils.almond_utils import (
     ISO_to_LANG,
@@ -288,39 +289,6 @@ class Paraphrase(NaturalSeq2Seq):
         answer = sentence  # means we calculate self-bleu
 
         return Example.from_raw(example_id, context, question, answer, preprocess=self.preprocess_field, lower=False)
-
-
-def inside_spans(start, spans):
-    for span in spans:
-        if span[0] <= start < span[1]:
-            return True
-    return False
-
-
-def return_sentences(text, regex_pattern, src_char_spans, lang):
-    sentences = []
-    cur = 0
-    for m in re.finditer(regex_pattern, text, flags=re.U):
-        if src_char_spans and not inside_spans(m.start(0), src_char_spans):
-            sentences.append(text[cur : m.start(0) + (1 if lang in ['zh', 'ja', 'ko'] else 0)])
-            cur = m.end(0)
-    if cur != len(text):
-        sentences.append(text[cur:])
-    return sentences
-
-
-def split_text_into_sentences(text, lang, src_char_spans):
-    if lang in ['en']:
-        sentences = return_sentences(text, '(?<!\w\.\w.)(?<![A-Z][a-z]\.)(?<=[\.!?])\s', src_char_spans, lang)
-    elif lang in ['zh', 'ja', 'ko']:
-        sentences = return_sentences(text, u'([!?。])\s?', src_char_spans, lang)
-    else:
-        import nltk
-
-        nltk.download('punkt', quiet=True)
-        sentences = nltk.sent_tokenize(text, language=ISO_to_LANG[lang])
-
-    return sentences
 
 
 @register_task('almond_translate')
